Switched from KInputDialog::getText() to KInputDialog::getInteger() for the Choose Game... action. Also cleaned up the internal handling of game numbers.
Certain parts of the code used ints while others used longs. This is a non issue on most 32bit system where ints and longs are the same size. However, on 64bit systems where they differ, it gave a false impression of range. Certain parts of the code are restricted to ints, so passing around longs was only confusing.

If there are any 64bit users out there who have already beaten the first 2147483647 games of Klondike, let me know and I'll consider upgrading the code to qint64.

svn path=/trunk/KDE/kdegames/kpat/; revision=982896

+++ dealer.h
@@ -115,8 +115,8 @@ public:
     // use this for autodrop times
     int speedUpTime( int delay ) const;
 
-    void setGameNumber(long gmn);
-    long gameNumber() const;
+    void setGameNumber(int gmn);
+    int gameNumber() const;
 
     int waiting() const { return _waiting; }
     void setWaiting(bool w);
@@ -191,7 +191,7 @@ public slots:
     void slotSolverFinished();
     void slotAutoDrop();
 
-    void startNew(long gameNumber = -1);
+    void startNew(int gameNumber = -1);
 
 protected slots:
     virtual void demo();
@@ -247,7 +247,7 @@ private:
     bool _autodrop;
     bool _usesolver;
     int _waiting;
-    long gamenumber;
+    int gamenumber;
 
     class DealerScenePrivate;
     DealerScenePrivate *d;
